feat: refactor PDF extraction methods to accept byte arrays and improve error handling
