+++ src/main/java/com/nexaworks/rafiq/service/PdfExtractorService.java
@@ -5,7 +5,8 @@ import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.concurrent.ExecutionException;
 
 public interface PdfExtractorService {
-    String extractPdf(MultipartFile pdfFile) throws IOException, DocumentException;
+    String extractPdf(MultipartFile pdfFile) throws IOException, DocumentException, ExecutionException, InterruptedException;
 }
