+++ src/main/java/com/nexaworks/rafiq/service/ServiceImpl/PdfExtractorServiceImpl.java
@@ -2,25 +2,21 @@ package com.nexaworks.rafiq.service.ServiceImpl;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
+import com.itextpdf.text.Document;
 import com.itextpdf.text.DocumentException;
+import com.itextpdf.text.Image;
+import com.itextpdf.text.pdf.PdfWriter;
 import com.nexaworks.rafiq.exception.custom.EmptyFileException;
 import com.nexaworks.rafiq.service.AiService;
 import com.nexaworks.rafiq.service.LabTestService;
 import com.nexaworks.rafiq.service.PdfExtractorService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import net.sourceforge.tess4j.ITesseract;
-import net.sourceforge.tess4j.Tesseract;
-import net.sourceforge.tess4j.TesseractException;
-import org.apache.pdfbox.pdmodel.PDDocument;
-import org.apache.pdfbox.rendering.PDFRenderer;
-import org.apache.pdfbox.text.PDFTextStripper;
+import org.apache.tomcat.util.http.fileupload.ByteArrayOutputStream;
 import org.springframework.ai.chat.client.ChatClient;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
-import java.awt.image.BufferedImage;
 import java.io.IOException;
-import java.io.InputStream;
 import java.util.Objects;
 import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
@@ -37,15 +33,34 @@ public class PdfExtractorServiceImpl implements PdfExtractorService {
     private final AiService aiService;
 
     @Override
-    public String extractPdf(MultipartFile pdfFile) throws IOException, DocumentException {
+    public String extractPdf(MultipartFile pdfFile) throws IOException, DocumentException, ExecutionException, InterruptedException {
         if (pdfFile.isEmpty()) {
             throw new EmptyFileException("The provided PDF file is empty. Please upload a valid file.");
         }
         CompletableFuture<UUID> testId = labTestService.saveTestPdf(pdfFile);
-        return aiService.extractLabResultsFromPdf(pdfFile);
+        byte[] pdfBytes = pdfFile.getBytes();
+        if (Objects.equals(pdfFile.getContentType(), "image/")) {
+            pdfBytes = convertImage(pdfBytes);
+        }
+        String result = aiService.extractLabResultsFromPdf(pdfBytes);
+        UUID id = testId.get();
+        ObjectNode jsonNode = (ObjectNode) new ObjectMapper().readTree(result);
+        jsonNode.put("testId", id.toString());
+        return  jsonNode.toString();
+    }
+    private static byte[] convertImage(byte[] pdfBytes) throws DocumentException, IOException {
+        ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
+        Document document = new Document();
+        PdfWriter.getInstance(document, outputStream);
+        document.open();
 
-
-
+        Image img = Image.getInstance(pdfBytes);
+        img.scaleToFit(document.getPageSize().getWidth(), document.getPageSize().getHeight());
+        img.setAlignment(Image.ALIGN_CENTER);
+        document.add(img);
+        document.close();
+        pdfBytes = outputStream.toByteArray();
+        return pdfBytes;
     }
